refactor common code for reporting serialevents
