+++ java/src/jmri/jmrix/AbstractSerialPortController.java
@@ -8,6 +8,8 @@ import purejavacomm.CommPortIdentifier;
 import purejavacomm.NoSuchPortException;
 import purejavacomm.PortInUseException;
 import purejavacomm.SerialPort;
+import purejavacomm.SerialPortEvent;
+import purejavacomm.SerialPortEventListener;
 
 /**
  * Provide an abstract base for *PortController classes.
@@ -207,6 +209,90 @@ abstract public class AbstractSerialPortController extends AbstractPortControlle
         return -1;
     }
 
+
+    /**
+     * Set event logging
+     */
+    protected void setPortEventLogging(SerialPort port) {
+        // arrange to notify later
+        try {
+            port.addEventListener(new SerialPortEventListener() {
+                @Override
+                public void serialEvent(SerialPortEvent e) {
+                    int type = e.getEventType();
+                    switch (type) {
+                        case SerialPortEvent.DATA_AVAILABLE:
+                            log.info("SerialEvent: DATA_AVAILABLE is " + e.getNewValue()); // NOI18N
+                            return;
+                        case SerialPortEvent.OUTPUT_BUFFER_EMPTY:
+                            log.info("SerialEvent: OUTPUT_BUFFER_EMPTY is " + e.getNewValue()); // NOI18N
+                            return;
+                        case SerialPortEvent.CTS:
+                            log.info("SerialEvent: CTS is " + e.getNewValue()); // NOI18N
+                            return;
+                        case SerialPortEvent.DSR:
+                            log.info("SerialEvent: DSR is " + e.getNewValue()); // NOI18N
+                            return;
+                        case SerialPortEvent.RI:
+                            log.info("SerialEvent: RI is " + e.getNewValue()); // NOI18N
+                            return;
+                        case SerialPortEvent.CD:
+                            log.info("SerialEvent: CD is " + e.getNewValue()); // NOI18N
+                            return;
+                        case SerialPortEvent.OE:
+                            log.info("SerialEvent: OE (overrun error) is " + e.getNewValue()); // NOI18N
+                            return;
+                        case SerialPortEvent.PE:
+                            log.info("SerialEvent: PE (parity error) is " + e.getNewValue()); // NOI18N
+                            return;
+                        case SerialPortEvent.FE:
+                            log.info("SerialEvent: FE (framing error) is " + e.getNewValue()); // NOI18N
+                            return;
+                        case SerialPortEvent.BI:
+                            log.info("SerialEvent: BI (break interrupt) is " + e.getNewValue()); // NOI18N
+                            return;
+                        default:
+                            log.info("SerialEvent of unknown type: " + type + " value: " + e.getNewValue()); // NOI18N
+                            return;
+                    }
+                }
+            }
+            );
+        } catch (java.util.TooManyListenersException ex) {
+            log.warn("cannot set listener for SerialPortEvents; was one already set?");
+        }
+        
+        try {
+            port.notifyOnFramingError(true);
+        } catch (Exception e) {
+            log.debug("Could not notifyOnFramingError: " + e); // NOI18N
+        }
+
+        try {
+            port.notifyOnBreakInterrupt(true);
+        } catch (Exception e) {
+            log.debug("Could not notifyOnBreakInterrupt: " + e); // NOI18N
+        }
+
+        try {
+            port.notifyOnParityError(true);
+        } catch (Exception e) {
+            log.debug("Could not notifyOnParityError: " + e); // NOI18N
+        }
+
+        try {
+            port.notifyOnOverrunError(true);
+        } catch (Exception e) {
+            log.debug("Could not notifyOnOverrunError: " + e); // NOI18N
+        }
+
+        port.notifyOnCarrierDetect(true);
+        port.notifyOnCTS(true);
+        port.notifyOnDSR(true);
+
+    }
+
+    
     Vector<String> portNameVector = null;
 
     /**
