+++ java/src/jmri/jmrix/zimo/mx1/Mx1Adapter.java
@@ -80,79 +80,18 @@ public class Mx1Adapter extends Mx1PortController implements jmri.jmrix.SerialPo
             }
             if (log.isDebugEnabled()) {
                 // report additional status
-                log.debug(" port flow control shows "
-                        + (activeSerialPort.getFlowControlMode() == SerialPort.FLOWCONTROL_RTSCTS_OUT ? "hardware flow control" : "no flow control"));
-            }
-            if (log.isDebugEnabled()) {
-                // arrange to notify later
-                activeSerialPort.addEventListener((SerialPortEvent e) -> {
-                    int type = e.getEventType();
-                    switch (type) {
-                        case SerialPortEvent.DATA_AVAILABLE:
-                            log.info("SerialEvent: DATA_AVAILABLE is " + e.getNewValue());
-                            return;
-                        case SerialPortEvent.OUTPUT_BUFFER_EMPTY:
-                            log.info("SerialEvent: OUTPUT_BUFFER_EMPTY is " + e.getNewValue());
-                            return;
-                        case SerialPortEvent.CTS:
-                            log.info("SerialEvent: CTS is " + e.getNewValue());
-                            return;
-                        case SerialPortEvent.DSR:
-                            log.info("SerialEvent: DSR is " + e.getNewValue());
-                            return;
-                        case SerialPortEvent.RI:
-                            log.info("SerialEvent: RI is " + e.getNewValue());
-                            return;
-                        case SerialPortEvent.CD:
-                            log.info("SerialEvent: CD is " + e.getNewValue());
-                            return;
-                        case SerialPortEvent.OE:
-                            log.info("SerialEvent: OE (overrun error) is " + e.getNewValue());
-                            return;
-                        case SerialPortEvent.PE:
-                            log.info("SerialEvent: PE (parity error) is " + e.getNewValue());
-                            return;
-                        case SerialPortEvent.FE:
-                            log.info("SerialEvent: FE (framing error) is " + e.getNewValue());
-                            return;
-                        case SerialPortEvent.BI:
-                            log.info("SerialEvent: BI (break interrupt) is " + e.getNewValue());
-                            return;
-                        default:
-                            log.info("SerialEvent of unknown type: " + type + " value: " + e.getNewValue());
-                    }
-                });
-                try {
-                    activeSerialPort.notifyOnFramingError(true);
-                } catch (Exception e) {
-                    log.debug("Could not notifyOnFramingError: " + e);
-                }
+                log.debug(" port flow control shows " // NOI18N
+                        + (activeSerialPort.getFlowControlMode() == SerialPort.FLOWCONTROL_RTSCTS_OUT ? "hardware flow control" : "no flow control")); // NOI18N
 
-                try {
-                    activeSerialPort.notifyOnBreakInterrupt(true);
-                } catch (Exception e) {
-                    log.debug("Could not notifyOnBreakInterrupt: " + e);
-                }
-
-                try {
-                    activeSerialPort.notifyOnParityError(true);
-                } catch (Exception e) {
-                    log.debug("Could not notifyOnParityError: " + e);
-                }
-
-                try {
-                    activeSerialPort.notifyOnOverrunError(true);
-                } catch (Exception e) {
-                    log.debug("Could not notifyOnOverrunError: " + e);
-                }
-
+                // log events
+                setPortEventLogging(activeSerialPort);
             }
 
             opened = true;
 
         } catch (NoSuchPortException p) {
             return handlePortNotFound(p, portName, log);
-        } catch (IOException | TooManyListenersException ex) {
+        } catch (IOException ex) {
             log.error("Unexpected exception while opening port {} trace follows: ", portName, ex);
             return "Unexpected error while opening port " + portName + ": " + ex;
         }
